🩺 The strictest Pint rules possible have been added

Closes HES-124

See merge request heseya/core!260

+++ app/Dtos/MetadataDto.php
@@ -8,11 +8,11 @@ use Heseya\Dto\Dto;
 class MetadataDto extends Dto
 {
     private string $name;
-    private bool|int|float|string|null $value;
+    private bool|float|int|string|null $value;
     private bool $public;
     private string $value_type;
 
-    public static function manualInit(string $name, bool|int|float|string|null $value, bool $public): self
+    public static function manualInit(string $name, bool|float|int|string|null $value, bool $public): self
     {
         return new self(
             name: $name,
@@ -27,7 +27,7 @@ class MetadataDto extends Dto
         return $this->name;
     }
 
-    public function getValue(): bool|int|float|string|null
+    public function getValue(): bool|float|int|string|null
     {
         return $this->value;
     }
