+++ app/Dtos/MetadataPersonalDto.php
@@ -8,10 +8,10 @@ use Heseya\Dto\Dto;
 class MetadataPersonalDto extends Dto
 {
     private string $name;
-    private bool|int|float|string|null $value;
+    private bool|float|int|string|null $value;
     private string $value_type;
 
-    public static function manualInit(string $name, bool|int|float|string|null $value): self
+    public static function manualInit(string $name, bool|float|int|string|null $value): self
     {
         return new self(
             name: $name,
@@ -25,7 +25,7 @@ class MetadataPersonalDto extends Dto
         return $this->name;
     }
 
-    public function getValue(): bool|int|float|string|null
+    public function getValue(): bool|float|int|string|null
     {
         return $this->value;
     }
